refactor: centralize validation utilities
